make configuration be working from main, instead of state::new and make state::new accept the configuration structure as a parameter
this makes it easier to make, for example, speech have the desired rate from startup, logging be initialised from the config with the desired filter and perhaps a path location, stuff like that
